adaption in train for linearRegressionBeam and WriteDBTask

+++ train/src/main/java/at/ac/uibk/dps/streamprocessingapplications/tasks/LinearRegressionTrainBatched.java
@@ -134,7 +134,7 @@ public class LinearRegressionTrainBatched extends AbstractTask {
       l.warn("error training decision tree", e);
     }
 
-    return Float.valueOf(1);
+    return 1F;
     //		return Float.valueOf(Float.MIN_VALUE);
   }
 }
